util 패키지 추가, Distance 클래스 추가, CustomerService 분리
UserService와 CustomerService를 분리했습니다. Distance 클래스의 distance 메소드를 사용하면 두 좌표의 거리값을 double 타입으로 반환합니다. (단위 : km)
