+++ src/main/java/zerobase/demo/user/service/Impl/UserServiceImpl.java
@@ -33,8 +33,6 @@ public class UserServiceImpl implements UserService {
 
 	private final MailComponents mailComponents;
 	private final UserRepository userRepository;
-	private final ReviewRepository reviewRepository;
-	private final OrderRepository orderRepository;
 
 	@Override
 	public boolean createUser(UserDto userDto) {
@@ -163,38 +161,6 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public boolean userAddReview(ReviewDto fromRequest, String userId) {
-		reviewRepository.findByOrderId(fromRequest.getOrderId())
-				.ifPresent(t -> {throw new UserException(ResponseCode.ALREADY_ADDED_REVIEW);});
-
-		Order order = orderRepository.findById(fromRequest.getOrderId())
-				.orElseThrow(() -> new UserException(ResponseCode.ORDER_NOT_FOUND));
-
-		if (!Objects.equals(order.getRestaurantId(), fromRequest.getRestaurantId())) {
-			throw new UserException(ResponseCode.DIFF_ORDER_ID);
-		}
-		if (order.getOrderTime().plusDays(3).isBefore(LocalDateTime.now())) {
-			throw new UserException(ResponseCode.TOO_OLD_AN_ORDER);
-		}
-		if (!order.getUserId().equals(userRepository.findByUserId(userId).get().getId())) {
-			throw new UserException(ResponseCode.NOT_MY_ORDER);
-		}
-
-		Review review = new Review().builder()
-				.writerId(userRepository.findByUserId(userId).get().getId())
-				.orderId(fromRequest.getOrderId())
-				.restaurantId(fromRequest.getRestaurantId())
-				.summary(fromRequest.getSummary())
-				.content(fromRequest.getContent())
-				.reviewAddTime(LocalDateTime.now())
-				.build();
-
-		reviewRepository.save(review);
-
-		return false;
-	}
-
-	@Override
 	public UserDetails loadUserByUsername(String userId) {
 
 		User user = userRepository.findByUserId(userId)
@@ -225,17 +191,6 @@ public class UserServiceImpl implements UserService {
 
 		return new org.springframework.security.core.userdetails.User(user.getUserId(),
 			user.getPassword(), grantedAuthorities);
-	}
-
-	@Override
-	public List<OrderDto> getMyOrderList(String userId) {
-
-		List<Order> orderDtoList = orderRepository
-				.findAllByUserId(userRepository.findByUserId(userId).get().getId());
-		if (orderDtoList.isEmpty()) throw new UserException(ResponseCode.THERE_IS_NO_ORDER);
-
-
-		return OrderDto.fromEntity(orderDtoList);
 	}
 
 }
